Remove the depenency on the npm phantomjs package
Medium’s phantomjs package installs a binary for phantomjs.
Unfortunately, it is in a holding pattern for supporting phantomjs 2.0
because linux binaries aren't readily available. See
https://github.com/ariya/phantomjs/issues/12948 and
https://github.com/Medium/phantomjs/issues/288.

Removing this dependency means that prerender will expect the
`phantomjs` binary to exist on your path, which is preferable because
it gives the end user flexibility on how to install phantomjs (i.e. via
a package manager, built from source, etc).